Correction exemple

+++ samples/C++17/22 - sample.cpp
@@ -8,7 +8,7 @@ int main()
 {
   std::string in = "abcdefgh";
   std::string out;
-  std::mt19937 gen{time(nullptr)};
+  std::mt19937 gen{static_cast<unsigned long>(time(nullptr))};
   std::sample(std::begin(in), std::end(in), std::back_inserter(out), 5, gen);
 
   for(auto it : out)
